refactor(elements): extract Scope+Locators sub-modules
Phase 12a §1: split CreateElementMediator.ts (2 483 LoC god module,
128 internal helpers) into `Create/*` sub-modules. This first sub-step
moves the form-scoping primitives and per-kind locator dispatch:

- `Create/Scope.ts` (39 LoC): `CLICK_RACE_TIMEOUT`, `NO_FORM_ANCHOR`,
  `LocatorContext` type, `applyFormScope` — the single point where
  form-membership becomes a deterministic Locator-chained filter.

- `Create/Locators.ts` (200 LoC): 10 per-kind locator builders +
  `LOCATOR_KIND_BUILDERS` OCP dispatch + `buildCandidateLocators` /
  `buildCandidateLocatorsBase` public API. Xpath helpers
  (`CLICK_ANCESTORS`, `XPATH_PREFIX_BY_SCOPE`, `relativizeXpath`)
  are private to this file — keeps Pipeline Rule #15 satisfied.

- `Create/index.ts`: explicit-symbol barrel routing both sub-modules
  PLUS `ActiveState` (transient pass-through; relocates to PhaseControls
  in Commit 6). The barrel keeps the façade under the 15-dependency cap
  during the multi-commit decomposition.

Façade shrinks 261 lines (2 483 → 2 222). Public surface is BYTE-IDENTICAL
(`lib/index.d.ts` + `lib/index.d.cts` SHA256 unchanged vs baseline).
All 30 Mediator/Elements test suites (270 tests) still pass.

No behavioural changes — pure structural decoupling.

Phase 12a sub-step 1/6: Scope + Locators
Closes part of v8.5.0 Phase 12a (KG-driven decoupling — split god module).
